move loadBallotImageUtil
