+++ libs/backend/src/index.ts
@@ -4,6 +4,7 @@ export * from './get_usb_drives';
 export * from './scan';
 export * from './split';
 export * from './list_directory';
+export * from './load_ballot_image';
 export * from './cast_vote_record_report_import';
 export * from './ballot_package';
 export * from './mock_usb';
